requirements.txt
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
numpy==1.23.0
opencv-python==4.6.0.66
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyserial==3.5
python-decouple==3.6
python-engineio==4.3.2
python-socketio==5.6.0
remio==0.1.1
requests==2.28.0
simplejpeg==1.6.4
urllib3==1.26.9
websocket-client==1.3.2
